pyproject.toml
[tool.poetry]
name = "kaggle"
version = "0.1.0"
description = ""
authors = [""]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
numpy = "^1.21.3"
pandas = "^1.3.4"
jupyter = "^1.0.0"
ipython = "^7.29.0"
mypy = "^0.910"
flake8 = "^4.0.1"
black = "^21.10b0"
python-dotenv = "^0.19.1"
torch = "^1.10.0"
pytorch-lightning = "^1.5.0"
tqdm = "^4.62.3"
matplotlib = "^3.4.3"
seaborn = "^0.11.2"
scikit-learn = "<0.24.2"
SoundFile = "^0.10.3"
librosa = "^0.8.1"
opencv-python = "^4.5.4"
efficientnet-pytorch = "^0.7.1"
torchmetrics = "^0.6.0"
Pillow = "^8.4.0"
torchvision = "^0.11.1"
torchinfo = "^1.5.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
